LayoutTests/platform/chromium-mac/plugins/keyboard-events-expected.txt

CONSOLE MESSAGE: line 18: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 18: PLUGIN: mouseDown at (12, 12)
CONSOLE MESSAGE: line 19: PLUGIN: mouseUp at (12, 12)
CONSOLE MESSAGE: line 22: PLUGIN: keyDown 'a'
CONSOLE MESSAGE: line 22: PLUGIN: keyUp 'a'
CONSOLE MESSAGE: line 23: PLUGIN: keyDown 'b'
CONSOLE MESSAGE: line 23: PLUGIN: keyUp 'b'
CONSOLE MESSAGE: line 24: PLUGIN: keyDown 'c'
CONSOLE MESSAGE: line 24: PLUGIN: keyUp 'c'

This test checks if a plug-in can receive keyboard events sent from eventSender. This is a test for Bug 34936.
